+++ src/main/java/org/janelia/saalfeldlab/n5/Bzip2BlockReaderWriter.java
@@ -26,40 +26,23 @@
 package org.janelia.saalfeldlab.n5;
 
 import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.nio.channels.ByteChannel;
-import java.nio.channels.Channels;
-import java.nio.channels.FileChannel;
+import java.io.InputStream;
+import java.io.OutputStream;
 
 import org.apache.commons.compress.compressors.bzip2.BZip2CompressorInputStream;
 import org.apache.commons.compress.compressors.bzip2.BZip2CompressorOutputStream;
-import org.apache.commons.io.IOUtils;
 
-public class Bzip2BlockReaderWriter implements BlockReader, BlockWriter {
+public class Bzip2BlockReaderWriter extends AbstractBlockReaderWriter {
 
 	@Override
-	public <T, B extends DataBlock<T>> void read(
-			final B dataBlock,
-			final ByteChannel channel) throws IOException {
+	protected InputStream getInputStream(final InputStream in) throws IOException {
 
-		final ByteBuffer buffer;
-		try (final BZip2CompressorInputStream in = new BZip2CompressorInputStream(Channels.newInputStream(channel))) {
-			final byte[] bytes = IOUtils.toByteArray(in);
-			buffer = ByteBuffer.wrap(bytes);
-		}
-		dataBlock.readData(buffer);
+		return new BZip2CompressorInputStream(in);
 	}
 
 	@Override
-	public <T> void write(
-			final DataBlock<T> dataBlock,
-			final FileChannel channel) throws IOException {
+	protected OutputStream getOutputStream(final OutputStream out) throws IOException {
 
-		final ByteBuffer buffer = dataBlock.toByteBuffer();
-		try (final BZip2CompressorOutputStream out = new BZip2CompressorOutputStream(Channels.newOutputStream(channel))) {
-			out.write(buffer.array());
-			out.finish();
-		}
-		channel.truncate(channel.position());
+		return new BZip2CompressorOutputStream(out);
 	}
 }
