+++ src/main/java/org/janelia/saalfeldlab/n5/GzipBlockReaderWriter.java
@@ -26,40 +26,23 @@
 package org.janelia.saalfeldlab.n5;
 
 import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.nio.channels.ByteChannel;
-import java.nio.channels.Channels;
-import java.nio.channels.FileChannel;
+import java.io.InputStream;
+import java.io.OutputStream;
 
 import org.apache.commons.compress.compressors.gzip.GzipCompressorInputStream;
 import org.apache.commons.compress.compressors.gzip.GzipCompressorOutputStream;
-import org.apache.commons.io.IOUtils;
 
-public class GzipBlockReaderWriter implements BlockReader, BlockWriter {
+public class GzipBlockReaderWriter extends AbstractBlockReaderWriter {
 
 	@Override
-	public <T, B extends DataBlock<T>> void read(
-			final B dataBlock,
-			final ByteChannel channel) throws IOException {
+	protected InputStream getInputStream(final InputStream in) throws IOException {
 
-		final ByteBuffer buffer;
-		try (final GzipCompressorInputStream in = new GzipCompressorInputStream(Channels.newInputStream(channel))) {
-			final byte[] bytes = IOUtils.toByteArray(in);
-			buffer = ByteBuffer.wrap(bytes);
-		}
-		dataBlock.readData(buffer);
+		return new GzipCompressorInputStream(in);
 	}
 
 	@Override
-	public <T> void write(
-			final DataBlock<T> dataBlock,
-			final FileChannel channel) throws IOException {
+	protected OutputStream getOutputStream(final OutputStream out) throws IOException {
 
-		final ByteBuffer buffer = dataBlock.toByteBuffer();
-		try (final GzipCompressorOutputStream os = new GzipCompressorOutputStream(Channels.newOutputStream(channel))) {
-			os.write(buffer.array());
-			os.flush();
-			channel.truncate(channel.position());
-		}
+		return new GzipCompressorOutputStream(out);
 	}
 }
